[bitnami/redis] Fix: avoid eager evaluation of ternary when resolving user secrets (#1472)
Helm’s ternary function (from Sprig) is eagerly evaluated, which
caused `include "common.secrets.get"` to run even when
`.Values.auth.acl.userSecret` was empty. This resulted in errors
when the referenced Secret did not exist.

Replaced ternary with an explicit `if` guard so that secret lookup
is only performed when a userSecret is actually set.

Signed-off-by: Sina Safari <sina.devel@gmail.com>
Co-authored-by: Sina Safari <sina.devel@gmail.com>